[#187] Refactor: 전반적인 예외 처리 방식 개선 (#277)
* feat: 각 도메인에서 최적화된 커스텀 예외를 발생시키는 BusinessException 클래스 구현

* feat: GlobalExceptionHandler 클래스 리팩토링

* feat: Enum 형태의 새 공통 ResponseCode 추가 후 GlobalExceptionHandler에 반영

* feat: Alarm 도메인에서 Enum 형태의 AlarmResponseCode 추가

* feat: Alarm 도메인의 Controller 클래스에서 AlarmResponseCode 적용

* feat: Alarm 도메인의 Service 클래스에서 AlarmResponseCode 적용

* refactor: 불필요한 코드 블록 삭제

* feat: Auth 도메인의 Controller 클래스에서 BusinessException과 AuthResponseCode 적용

* feat: Auth 도메인의 Service 클래스에서 BusinessException과 AuthResponseCode 적용

* feat: Auth 도메인의 Controller 클래스에서 BusinessException과 AuthResponseCode 적용

* feat: OAuth 도메인의 Controller 클래스에서 BusinessException과 OAuthResponseCode 적용

* feat: OAuth 도메인의 service 클래스에서 BusinessException과 OAuthResponseCode 적용

* feat: Auth/OAuth 도메인의 에외처리가 필요한 다른 로직에서 새로운 방식으로 수정

* feat: Auth/OAuth 의 ResponseCode enum 추가

* feat: 즁복되는 예외 메서드 정리

* feat: Auth 도메인 내에서 기존 예외처리 로직 삭제

* feat: user 도메인의 ResponseCode enum 추가

* feat: 모든 user not found 예외를 user 도메인의 ResponseCode enum으로 수정

* feat: 모든 user not found 예외의 응답 메세지를 커스터마이징

* feat: user 도메인에서 예외처리 방식 개선을 위한 UserResponseCode enum 추가

* feat: user 도메인의 controller와 service에서 UserResponseCode을 적용

* feat: interests 도메인의 controller와 service에서 UserResponseCode을 적용

* feat: interests 도메인의 InterestsResponseCode을 추가

* feat: 공통 예외 ResponseCode에 AI 서버 예외 추가

* delete: interests 도메인의 기존 예외 클라스 삭제

* feat: BusinessException으로 수정

* feat: channel 도메인의 ResponseCode enum 추가

* feat: channel 도메인의 ResponseCode enum 추가

* feat: channel 도메인의 Controller와 Service 클래스에 ResponseCode enum 적용

* feat: channel 도메인의 AsyncChannelService와 SseChannelService 클래스에 ResponseCode enum 적용

* feat: 안 쓰는 import 삭제

* refactor: 각 도메인 패키지내에 있는 exception 패키지 이름을 responsecode 로 수정

* refactor: SSE 로직에세 SSE 전용 예외 추가

* WIP: 임시 작업 저장

+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/alarm/service/AlarmService.java
@@ -15,10 +15,11 @@ import com.hertz.hertz_be.domain.alarm.repository.AlarmRepository;
 import com.hertz.hertz_be.domain.alarm.repository.UserAlarmRepository;
 import com.hertz.hertz_be.domain.channel.entity.SignalRoom;
 import com.hertz.hertz_be.domain.channel.entity.enums.MatchingStatus;
-import com.hertz.hertz_be.domain.channel.exception.UserNotFoundException;
 import com.hertz.hertz_be.domain.user.entity.User;
+import com.hertz.hertz_be.domain.user.responsecode.UserResponseCode;
 import com.hertz.hertz_be.domain.user.repository.UserRepository;
-import com.hertz.hertz_be.global.exception.InternalServerErrorException;
+import com.hertz.hertz_be.global.common.NewResponseCode;
+import com.hertz.hertz_be.global.exception.BusinessException;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.data.domain.Page;
@@ -27,7 +28,6 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.support.TransactionSynchronization;
-import org.springframework.transaction.support.TransactionSynchronizationAdapter;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import java.time.LocalDateTime;
@@ -52,7 +52,11 @@ public class AlarmService {
     @Transactional
     public void createNotifyAlarm(CreateNotifyAlarmRequestDto dto, Long userId) {
         User notifyWriter = userRepository.findById(userId)
-                .orElseThrow(UserNotFoundException::new);
+                .orElseThrow(() -> new BusinessException(
+                        UserResponseCode.USER_NOT_FOUND.getCode(),
+                        UserResponseCode.USER_NOT_FOUND.getHttpStatus(),
+                        "공지 알람 작성 API를 요청한 사용자가 존재하지 않습니다."
+                ));
 
         AlarmNotification alarmNotification = AlarmNotification.builder()
                 .title(dto.getTitle())
@@ -181,7 +185,11 @@ public class AlarmService {
                                 report.getCreatedAt().toString()
                         );
                     } else {
-                        throw new InternalServerErrorException();
+                        throw new BusinessException(
+                                NewResponseCode.INTERNAL_SERVER_ERROR.getCode(),
+                                NewResponseCode.INTERNAL_SERVER_ERROR.getHttpStatus(),
+                                "알람 리스트 반환 중 예외 발생했습니다."
+                        );
                     }
                 })
                 .collect(Collectors.toList());
@@ -205,8 +213,13 @@ public class AlarmService {
 
     @Transactional
     public void deleteAlarm(Long alarmId, Long userId) {
-        userRepository.findById(userId)
-                .orElseThrow(UserNotFoundException::new);
+        if (!userRepository.existsById(userId)) {
+            throw new BusinessException(
+                    UserResponseCode.USER_NOT_FOUND.getCode(),
+                    UserResponseCode.USER_NOT_FOUND.getHttpStatus(),
+                    "알람 삭제 API를 요청한 사용자가 존재하지 않습니다."
+            );
+        }
 
         alarmRepository.deleteById(alarmId);
     }
